+++ acoustics/acoustics_data_record_auv/acoustics_data_record_auv/acoustics_data_record_lib.py
@@ -8,7 +8,7 @@ class AcousticsDataRecordLib:
     def __init__(self, ros2_package_directory: str = "") -> None:
         # Global variables for .csv file manipulation ----------
         # Get the path for the directory where we will store our data
-        self.acoustics_data_directory = ros2_package_directory + "acoustics_data/"
+        self.acoustics_data_directory = ros2_package_directory + "/acoustics_data/"
 
         timestamp = time.strftime("%Y-%m-%d_%H:%M:%S")
         data_file_name = "acoustics_data_" + timestamp + ".csv"
