+++ acoustics/acoustics_data_record_auv/launch/acoustics_data_record_launch.py
@@ -18,23 +18,15 @@ def generate_launch_description() -> LaunchDescription:
         for acoustics_data_record.
 
     """
-    # Path to the YAML file
-    yaml_file_path = os.path.join(
-        get_package_share_directory("acoustics_interface"),
-        "../../../../",  # Go to the workspace
-        "src/vortex-auv/auv_setup/config/robots/",  # Go inside where yamal files are located at
-        "orca.yaml",
-    )
 
     return LaunchDescription(
         [
             Node(
-                package="acoustics_data_record",
-                namespace="acoustics_data_record",
+                package="acoustics_data_record_auv",
                 executable="acoustics_data_record_node.py",
-                name="acoustics_data_record_node",
+                name="acoustics_data_record_auv_node",
                 output="screen",
-                parameters=[yaml_file_path],
+                parameters=[],
             ),
         ]
     )
