+++ acoustics/acoustics_data_record_auv/package.xml
@@ -1,18 +1,14 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>acoustics_data_record</name>
+  <name>acoustics_data_record_auv</name>
   <version>1.0.0</version>
   <description>Records data from the acoustics</description>
   <maintainer email="karate.martynas1@gmail.com">vortex</maintainer>
   <license>MIT</license>
 
-  <!-- Execution dependencies -->
-  <exec_depend>rclpy</exec_depend>
-  <exec_depend>std_msgs</exec_depend>
-  <exec_depend>ros2launch</exec_depend>
+  <buildtool_depend>ament_cmake</buildtool_depend>
 
-  <!-- Export the build type as ament_cmake for CMake projects -->
   <export>
     <build_type>ament_cmake</build_type>
   </export>
